refactor:removed the search from base repository and implemented it separately in each repository
