+++ app/Repositories/CategoryRepository.php
@@ -5,10 +5,9 @@ namespace App\Repositories;
 use App\Models\Category;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Pagination\LengthAwarePaginator;
 
-/**
- * @method LengthAwarePaginator search(array $filters = [], int $perPage, int $page)
- */
+
 class CategoryRepository extends BaseRepository
 {
     public function __construct(Category $category)
@@ -26,5 +25,23 @@ class CategoryRepository extends BaseRepository
             $category->update($data);
             Log::info('Category updated', ['category_id' => $category->id]);
         });
+    }
+    public function search(array $filters = [], int $perPage, int $page): LengthAwarePaginator
+    {
+        unset($filters['per_page'], $filters['page']);
+
+        $query = $this->model->query();
+
+        foreach ($filters as $column => $value) {
+            if ($value !== null) {
+                if (in_array($column, ['parent_id'])) {
+                    $query->where($column, $value);
+                } else {
+                    $query->where($column, 'LIKE', "%{$value}%");
+                }
+            }
+        }
+
+        return $query->paginate($perPage, ['*'], 'page', $page);
     }
 }
