+++ app/Repositories/NoteRepository.php
@@ -5,10 +5,9 @@ namespace App\Repositories;
 use App\Models\Note;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Pagination\LengthAwarePaginator;
 
-/**
- * @method LengthAwarePaginator search(array $filters = [], int $perPage, int $page)
- */
+
 class NoteRepository extends BaseRepository
 {
     public function __construct(Note $note)
@@ -26,5 +25,23 @@ class NoteRepository extends BaseRepository
             $note->update($data);
             Log::info('Note updated', ['note_id' => $note->id]);
         });
+    }
+    public function search(array $filters = [], int $perPage, int $page): LengthAwarePaginator
+    {
+        unset($filters['per_page'], $filters['page']);
+
+        $query = $this->model->query();
+
+        foreach ($filters as $column => $value) {
+            if ($value !== null) {
+                if (in_array($column, ['category_id'])) {
+                    $query->where($column, $value);
+                } else {
+                    $query->where($column, 'LIKE', "%{$value}%");
+                }
+            }
+        }
+
+        return $query->paginate($perPage, ['*'], 'page', $page);
     }
 }
